requirements.txt
yfinance==0.2.51
pandas==2.2.3
requests==2.32.3
anthropic==0.40.0
PyYAML==6.0.2
pytz==2024.2
python-dotenv==1.0.1
